+++ models/env.src
@@ -1,7 +1,6 @@
 # Use the src/build directory
 export SRC_DIR=${PWD}/..
 export BUILD_DIR=$SRC_DIR/build
-export MODELS_SRC=/models
 export MODELS_ROOT=/models
 
 export HALO_BIN=$BUILD_DIR/bin/halo
