+++ models/vision/detection/yolo/demo.cc
@@ -1,4 +1,4 @@
-//===- demo.cc --------------------------------------------------------===//
+//===- demo.cc ------------------------------------------------------------===//
 //
 // Copyright (C) 2019-2020 Alibaba Group Holding Limited.
 //
